+++ src/cli.rs
@@ -1,20 +1,6 @@
-use crate::cmds;
+use crate::cmds::SubCommand;
 use crate::config::Config;
 use clap::Parser;
-
-#[derive(clap::Parser, Debug)]
-pub enum SubCommand {
-    #[clap(
-        name = "assume-role",
-        about = "Temporary credentials for an assumed AWS IAM Role"
-    )]
-    AssumeRole(cmds::AssumeRole),
-    #[clap(
-        name = "session-token",
-        about = "Temporary credentials for an AWS IAM user"
-    )]
-    GetSessionToken(cmds::SessionToken),
-}
 
 #[derive(clap::Parser, Debug)]
 #[command(author, version, about, long_about = None)]
